requirements.txt
langchain
langgraph
langchain-core
langchain-openai
langchain-groq
langchain-community

pydantic
pydantic-settings

langgraph-checkpoint-sqlite

fastapi
uvicorn
python-dotenv

nltk
ipython
